Audio: sanitize `volume`
* move bound checkes to platform-independent layer
* snap to 0 / 1 to use memcpy/memset when close to those factors
* use float instead of qreal in the private volume APIs consistently
* use epsilon sufficient for 22bit resolution

Pick-to: 6.8
(cherry picked from commit e8312bc5b8677fff207e44eb4b324e5bd5845245)

+++ src/multimedia/darwin/qdarwinaudiosink_p.h
@@ -51,8 +51,8 @@ public:
     qint64 processedUSecs() const override;
     QAudioFormat format() const override;
 
-    void setVolume(qreal volume) override;
-    qreal volume() const override;
+    void setVolume(float volume) override;
+    float volume() const override;
 
 private:
     void resumeStreamIfNecessary();
@@ -62,7 +62,7 @@ private:
     const QAudioFormat m_audioFormat;
     std::optional<int> m_internalBufferSize;
 
-    qreal m_volume = 1.;
+    float m_volume = 1.f;
 
     std::shared_ptr<QCoreAudioSinkStream> m_stream;
 };
