+++ src/multimedia/pulseaudio/qpulseaudiosink_p.h
@@ -55,8 +55,8 @@ public:
     void setFormat(const QAudioFormat &format);
     QAudioFormat format() const override;
 
-    void setVolume(qreal volume) override;
-    qreal volume() const override;
+    void setVolume(float volume) override;
+    float volume() const override;
 
     void streamUnderflowCallback();
     void streamDrainedCallback();
@@ -101,7 +101,7 @@ private:
     qint64 m_elapsedTimeOffset = 0;
     mutable qint64 averageLatency = 0; // average latency
     mutable qint64 lastProcessedUSecs = 0;
-    qreal m_volume = 1.0;
+    float m_volume = 1.0;
 
     std::atomic<pa_operation *> m_drainOperation = nullptr;
     qsizetype m_bufferSize = 0;
